+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(iobed.shiny)
+
+test_check("iobed.shiny")