+++ NEWS.md
@@ -1,3 +1,4 @@
 # iobed.shiny 0.0.0.9000
 
+* Completed `{golem}`'s `dev/01-start.R`.
 * Added a `NEWS.md` file to track changes to the package.